% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Daily_Popup.R
\name{Daily_Popup}
\alias{Daily_Popup}
\title{Creates the Daily Popup Data Table}
\usage{
Daily_Popup(temp_table)
}
\arguments{
\item{temp_table}{The table to generate the report.}
}
\value{
dt_report A Daily Popup Data Table.
}
\description{
This function creates a Daily Popup Data Table.
}
\details{
Requires: DT
}
\examples{

}
