% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inv_Bin_Detail.R
\name{Inv_Bin_Detail}
\alias{Inv_Bin_Detail}
\title{Creates the Inventory Bin Detail Data Table}
\usage{
Inv_Bin_Detail(b_inv_table)
}
\arguments{
\item{b_inv_table}{The Inventory Table for the user.}
}
\value{
A Inventory Bin Detail Data Table.
}
\description{
This function creates a Inventory Bin Detail Data Table.
}
\details{
Requires: DT
}
\examples{

}
