% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inv_Popup.R
\name{Inv_Popup}
\alias{Inv_Popup}
\title{Creates the Inventory Popup Data Table}
\usage{
Inv_Popup(temp_table)
}
\arguments{
\item{temp_table}{The table to generate the report.}
}
\value{
dt_report An Inventory Popup Data Table.
}
\description{
This function creates a Inventory Popup Data Table.
}
\details{
Requires: DT
}
\examples{

}
