% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inventory_Table.R
\name{Inventory_Table}
\alias{Inventory_Table}
\title{Creates the Inventory Detail Data Table}
\usage{
Inventory_Table(inv_table, rem_flag = FALSE)
}
\arguments{
\item{inv_table}{The inventory table to generate the report.}

\item{rem_flag}{The flag to indicate whether the Edit Button should be removed from output.}
}
\value{
An Inventory Detail Data Table.
}
\description{
This function creates a Inventory Detail Data Table.
}
\details{
Requires: DT, data.table
}
\examples{

}
