% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sales_Activity_Table.R
\name{Sales_Activity_Table}
\alias{Sales_Activity_Table}
\title{Creates the Sales Activity Data Table}
\usage{
Sales_Activity_Table(sales_activity)
}
\arguments{
\item{sales_activity}{The table to be used for the report.}
}
\value{
sar_DT A Sales Activity Data Table.
}
\description{
This function creates a Sales Activity Data Table.
}
\details{
Requires: DT, dplyr
}
\examples{

}
