% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sales_Popup.R
\name{Sales_Popup}
\alias{Sales_Popup}
\title{Creates the Sales Popup Data Table}
\usage{
Sales_Popup(temp_table)
}
\arguments{
\item{temp_table}{The table to be used for the report.}
}
\value{
dt_report A Sales Popup Data Table.
}
\description{
This function creates a Sales Popup Data Table.
}
\details{
Requires: DT, dplyr
}
\examples{

}
